pyproject.toml
[project]
name = "travel-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Supachai_280246",email = "6510110469@email.psu.ac.th"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[dotenv] (>=2.11.7,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

